backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pytesseract
Pillow
pdfplumber
Unidecode